New Package Implemented
The reason for the separate package is that when I first started this project, I forgot to add it to GitHub. As a result, I ended up creating two separate projects: one for testing my code and animations, which is not connected to GitHub, and another as the main project that is connected to GitHub. Initially, I thought I could manually transfer my code between the two projects, but I realized it would be a waste of time. So instead, I implemented it as a package.

P.S.: I considered working on two different scenes within the same project, but I was worried that it might cause issues with GitHub.