RayDistance , Calculations and enviromental changes
I added functionality in the base class AiFiniteStates to allow dynamic adjustment of the ray distance. This includes methods to set and reset the ray distance, using Invoke method to schedule a reset back to the original distance after a specified duration. When the AI enters the alerted state,  SetRayDistance to increase the ray distance. This is crucial to give the AI a temporary boost in its sensory capabilities while in alerted state, which can be critical for gameplay mechanics where the AI needs to detect threats or targets more effectively during high-alert phases.

----------------------------------------

In GameManager script to include a method that calculates the distance between two AI entities (Ai_a and Ai_b). This is beneficial because the GameManager holds references to these entities, making it a central point from which to manage and access game-wide behaviors and data.The method I added, CalculateDistanceBetweenAIs(), utilizes the Unity method Vector3.Distance to find the straight-line distance between the positions of Ai_a and Ai_b. If either of the AI references is null, it logs an error and returns -1 to indicate failure in calculation.In future I`ll use this calculations to increase or decrease raycast thus I do not need a predefined variable for increase of raycast it can calculate automatically for best outcome.